modified and added customized graph class, naive tests passed

+++ src/com/lenss/yzeng/utils/Sortable.java
@@ -1,5 +1,5 @@
 package com.lenss.yzeng.utils;
 
 public interface Sortable{
-	public int getKey();
+	public double getKey();
 }
